refactor: move auth to libs/
